# docs/swagger.yaml
basePath: /v1
definitions:
...
  api.UpdateRetentionRequest:
    properties:
      alert_logs_days:
        type: integer
      audit_logs_days:
        type: integer
      system_event_logs_days:
        type: integer
      telemetry_influx_days:
        type: integer
    type: object
...
  models.SiteRetentionPolicy:
    properties:
      alert_logs_days:
        type: integer
      audit_logs_days:
        type: integer
      id:
        type: string
      site_id:
        type: string
      system_event_logs_days:
        type: integer
      telemetry_influx_days:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  response.Meta:
    properties:
      pagination:
        allOf:
        - $ref: '#/definitions/response.Pagination'
        description: Chỉ xuất hiện khi có phân trang, nếu không có sẽ tự động ẩn đi
      timestamp:
        type: string
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Linh Community
    url: https://github.com/linhcommunity/nlscada-cloud
  description: Backend API for NL SCADA Cloud
  title: NL SCADA Center API
  version: "1.0"
paths:
...
  /sites/{siteID}/config:
    get:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cấu hình đầy đủ
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Cấu hình tổng hợp cho Gateway
      tags:
      - Site Config
...
  /sites/{siteID}/retention-policies:
    get:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chính sách hiện tại
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SiteRetentionPolicy'
              type: object
      security:
      - BearerAuth: []
      summary: Xem chính sách xóa dữ liệu
      tags:
      - Retention
    put:
      consumes:
      - application/json
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Số ngày lưu trữ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateRetentionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Chính sách hiện tại
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SiteRetentionPolicy'
              type: object
      security:
      - BearerAuth: []
      summary: Cập nhật chính sách xóa dữ liệu
      tags:
      - Retention
  /sites/{siteID}/retention-policies/trigger-manual:
    post:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      responses:
        "202":
          description: Đã kích hoạt dọn dẹp
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Kích hoạt dọn dẹp ngay
      tags:
      - Retention
...
securityDefinitions:
  BearerAuth:
    description: Sử dụng "Bearer <token>", nhưng cookie session_token cũng được chấp
      nhận khi gọi từ browser
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"